// projd/ClientApp/src/app/models/Timesheet.ts
export class Timesheet {
  timesheetid: number;
  employeeid: number;
  managerid: number;
  startdate: string;
  enddate: string;
  dateapproved: string;
  day1: number;
  day2: number;
  day3: number;
  day4: number;
  day5: number;
  day6: number;
  day7: number;
  overtime: number;
  tstatus: number;
  comments: string;
  jwt: string;
}